pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "python-m2m-auth-demo"
version = "0.1.0"
description = "Python FastAPI server that demonstrates Wristband machine-to-machine (M2M) authentication."
authors = [
    {name = "Wristband", email = "support@wristband.dev"},
]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.133.0,<0.134.0",
    "httpx>=0.28.1,<0.29.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "uvicorn>=0.34.0,<0.35.0",
    "wristband-python-m2m-auth==0.1.0",
    "wristband-python-jwt==0.1.1",
]

[project.optional-dependencies]
dev = [
    "flake8>=6.0.0,<7.0.0",
    "flake8-pyproject>=1.2.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.10.0",
    "pip-audit>=2.0.0",
    "bandit>=1.7.0",
]

[project.urls]
"Homepage" = "https://wristband.dev"
"Repository" = "https://github.com/wristband-dev/m2m-python-auth-demo"
"Documentation" = "https://docs.wristband.dev"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/"]
mypy_path = ["src"]
namespace_packages = true
explicit_package_bases = true 

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 120
